back-end/api-1/requirements.txt
annotated-types==0.6.0
anyio==4.3.0
asn1crypto==1.5.1
boto3==1.34.72
botocore==1.34.72
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.5
defusedxml==0.8.0rc2
ecdsa==0.18.0
exceptiongroup==1.2.0
fastapi==0.110.0
filelock==3.13.3
future==0.18.3
greenlet==3.0.3
h11==0.14.0
idna==3.6
jmespath==1.0.1
jose==1.0.0
oauthlib==3.2.2
packaging==24.0
platformdirs==3.11.0
pyasn1==0.6.0
pycparser==2.21
pydantic==2.6.4
pydantic_core==2.16.3
PyJWT==2.8.0
pyOpenSSL==24.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.0.0
python-magic
python-multipart==0.0.9
python3-openid==3.2.0
pytz==2024.1
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9
s3transfer==0.10.1
six==1.16.0
sniffio==1.3.1
snowflake-connector-python==3.7.1
snowflake-sqlalchemy==1.5.1
social-auth-core==4.5.3
sortedcontainers==2.4.0
SQLAlchemy==1.4.52
starlette==0.36.3
tomlkit==0.12.4
typing_extensions==4.10.0
urllib3==1.26.18
uvicorn==0.29.0
